First part of refactor, including some of the config and most of the tokens actions
