+++ chains/solana/contracts/programs/ccip-router/src/instructions/mod.rs
@@ -1,0 +1,1 @@
+pub mod v1;
