+++ chains/solana/contracts/programs/ccip-router/src/lib.rs
@@ -33,6 +33,8 @@ use crate::ocr3base::*;
 mod fee_quoter;
 use crate::fee_quoter::*;
 
+mod instructions;
+
 // Anchor discriminators for CPI calls
 const CCIP_RECEIVE_DISCRIMINATOR: [u8; 8] = [0x0b, 0xf4, 0x09, 0xf9, 0x2c, 0x53, 0x2f, 0xf5]; // ccip_receive
 const TOKENPOOL_LOCK_OR_BURN_DISCRIMINATOR: [u8; 8] =
@@ -109,13 +111,7 @@ pub mod ccip_router {
         ctx: Context<TransferOwnership>,
         proposed_owner: Pubkey,
     ) -> Result<()> {
-        let mut config = ctx.accounts.config.load_mut()?;
-        require!(
-            proposed_owner != config.owner,
-            CcipRouterError::InvalidInputs
-        );
-        config.proposed_owner = proposed_owner;
-        Ok(())
+        instructions::v1::transfer_ownership(ctx, proposed_owner)
     }
 
     /// Accepts the ownership of the router by the proposed owner.
@@ -127,10 +123,7 @@ pub mod ccip_router {
     /// * `ctx` - The context containing the accounts required for accepting ownership.
     /// The new owner must be a signer of the transaction.
     pub fn accept_ownership(ctx: Context<AcceptOwnership>) -> Result<()> {
-        let mut config = ctx.accounts.config.load_mut()?;
-        config.owner = std::mem::take(&mut config.proposed_owner);
-        config.proposed_owner = Pubkey::new_from_array([0; 32]);
-        Ok(())
+        instructions::v1::accept_ownership(ctx)
     }
 
     /// Updates the fee aggregator in the router configuration.
@@ -144,9 +137,7 @@ pub mod ccip_router {
         ctx: Context<UpdateConfigCCIPRouter>,
         fee_aggregator: Pubkey,
     ) -> Result<()> {
-        let mut config = ctx.accounts.config.load_mut()?;
-        config.fee_aggregator = fee_aggregator;
-        Ok(())
+        instructions::v1::update_fee_aggregator(ctx, fee_aggregator)
     }
 
     /// Adds a new chain selector to the router.
@@ -167,38 +158,12 @@ pub mod ccip_router {
         source_chain_config: SourceChainConfig,
         dest_chain_config: DestChainConfig,
     ) -> Result<()> {
-        // Set source chain config & state
-        let source_chain_state = &mut ctx.accounts.source_chain_state;
-        validate_source_chain_config(new_chain_selector, &source_chain_config)?;
-        source_chain_state.version = 1;
-        source_chain_state.chain_selector = new_chain_selector;
-        source_chain_state.config = source_chain_config.clone();
-        source_chain_state.state = SourceChainState { min_seq_nr: 1 };
-
-        // Set dest chain config & state
-        let dest_chain_state = &mut ctx.accounts.dest_chain_state;
-        validate_dest_chain_config(new_chain_selector, &dest_chain_config)?;
-        dest_chain_state.version = 1;
-        dest_chain_state.chain_selector = new_chain_selector;
-        dest_chain_state.config = dest_chain_config.clone();
-        dest_chain_state.state = DestChainState {
-            sequence_number: 0,
-            usd_per_unit_gas: TimestampedPackedU224 {
-                value: [0; 28],
-                timestamp: 0,
-            },
-        };
-
-        emit!(SourceChainAdded {
-            source_chain_selector: new_chain_selector,
+        instructions::v1::add_chain_selector(
+            ctx,
+            new_chain_selector,
             source_chain_config,
-        });
-        emit!(DestChainAdded {
-            dest_chain_selector: new_chain_selector,
             dest_chain_config,
-        });
-
-        Ok(())
+        )
     }
 
     /// Disables the source chain selector.
@@ -213,16 +178,7 @@ pub mod ccip_router {
         ctx: Context<UpdateSourceChainSelectorConfig>,
         source_chain_selector: u64,
     ) -> Result<()> {
-        let chain_state = &mut ctx.accounts.source_chain_state;
-
-        chain_state.config.is_enabled = false;
-
-        emit!(SourceChainConfigUpdated {
-            source_chain_selector,
-            source_chain_config: chain_state.config.clone(),
-        });
-
-        Ok(())
+        instructions::v1::disable_source_chain_selector(ctx, source_chain_selector)
     }
 
     /// Disables the destination chain selector.
@@ -237,16 +193,7 @@ pub mod ccip_router {
         ctx: Context<UpdateDestChainSelectorConfig>,
         dest_chain_selector: u64,
     ) -> Result<()> {
-        let chain_state = &mut ctx.accounts.dest_chain_state;
-
-        chain_state.config.is_enabled = false;
-
-        emit!(DestChainConfigUpdated {
-            dest_chain_selector,
-            dest_chain_config: chain_state.config.clone(),
-        });
-
-        Ok(())
+        instructions::v1::disable_dest_chain_selector(ctx, dest_chain_selector)
     }
 
     /// Updates the configuration of the source chain selector.
@@ -263,15 +210,11 @@ pub mod ccip_router {
         source_chain_selector: u64,
         source_chain_config: SourceChainConfig,
     ) -> Result<()> {
-        validate_source_chain_config(source_chain_selector, &source_chain_config)?;
-
-        ctx.accounts.source_chain_state.config = source_chain_config.clone();
-
-        emit!(SourceChainConfigUpdated {
+        instructions::v1::update_source_chain_config(
+            ctx,
             source_chain_selector,
             source_chain_config,
-        });
-        Ok(())
+        )
     }
 
     /// Updates the configuration of the destination chain selector.
@@ -288,15 +231,7 @@ pub mod ccip_router {
         dest_chain_selector: u64,
         dest_chain_config: DestChainConfig,
     ) -> Result<()> {
-        validate_dest_chain_config(dest_chain_selector, &dest_chain_config)?;
-
-        ctx.accounts.dest_chain_state.config = dest_chain_config.clone();
-
-        emit!(DestChainConfigUpdated {
-            dest_chain_selector,
-            dest_chain_config,
-        });
-        Ok(())
+        instructions::v1::update_dest_chain_config(ctx, dest_chain_selector, dest_chain_config)
     }
 
     /// Updates the Solana chain selector in the router configuration.
@@ -311,11 +246,7 @@ pub mod ccip_router {
         ctx: Context<UpdateConfigCCIPRouter>,
         new_chain_selector: u64,
     ) -> Result<()> {
-        let mut config = ctx.accounts.config.load_mut()?;
-
-        config.solana_chain_selector = new_chain_selector;
-
-        Ok(())
+        instructions::v1::update_solana_chain_selector(ctx, new_chain_selector)
     }
 
     /// Updates the default gas limit in the router configuration.
@@ -331,11 +262,7 @@ pub mod ccip_router {
         ctx: Context<UpdateConfigCCIPRouter>,
         new_gas_limit: u128,
     ) -> Result<()> {
-        let mut config = ctx.accounts.config.load_mut()?;
-
-        config.default_gas_limit = new_gas_limit;
-
-        Ok(())
+        instructions::v1::update_default_gas_limit(ctx, new_gas_limit)
     }
 
     /// Updates the default setting for allowing out-of-order execution for other destination chains.
@@ -349,15 +276,10 @@ pub mod ccip_router {
         ctx: Context<UpdateConfigCCIPRouter>,
         new_allow_out_of_order_execution: bool,
     ) -> Result<()> {
-        let mut config = ctx.accounts.config.load_mut()?;
-
-        let mut v = 0_u8;
-        if new_allow_out_of_order_execution {
-            v = 1;
-        }
-        config.default_allow_out_of_order_execution = v;
-
-        Ok(())
+        instructions::v1::update_default_allow_out_of_order_execution(
+            ctx,
+            new_allow_out_of_order_execution,
+        )
     }
 
     /// Updates the minimum amount of time required between a message being committed and when it can be manually executed.
@@ -373,11 +295,10 @@ pub mod ccip_router {
         ctx: Context<UpdateConfigCCIPRouter>,
         new_enable_manual_execution_after: i64,
     ) -> Result<()> {
-        let mut config = ctx.accounts.config.load_mut()?;
-
-        config.enable_manual_execution_after = new_enable_manual_execution_after;
-
-        Ok(())
+        instructions::v1::update_enable_manual_execution_after(
+            ctx,
+            new_enable_manual_execution_after,
+        )
     }
 
     /// Registers the Token Admin Registry via the CCIP Admin
@@ -392,18 +313,11 @@ pub mod ccip_router {
         mint: Pubkey, // should we validate that this is a real token program?
         token_admin_registry_admin: Pubkey,
     ) -> Result<()> {
-        let token_admin_registry = &mut ctx.accounts.token_admin_registry;
-        token_admin_registry.version = 1;
-        token_admin_registry.administrator = token_admin_registry_admin;
-        token_admin_registry.pending_administrator = Pubkey::new_from_array([0; 32]);
-        token_admin_registry.lookup_table = Pubkey::new_from_array([0; 32]);
-
-        emit!(AdministratorRegistered {
-            token_mint: mint,
-            administrator: token_admin_registry_admin,
-        });
-
-        Ok(())
+        instructions::v1::register_token_admin_registry_via_get_ccip_admin(
+            ctx,
+            mint,
+            token_admin_registry_admin,
+        )
     }
 
     /// Registers the Token Admin Registry via the token owner.
@@ -416,28 +330,7 @@ pub mod ccip_router {
     pub fn register_token_admin_registry_via_owner(
         ctx: Context<RegisterTokenAdminRegistryViaOwner>,
     ) -> Result<()> {
-        let token_mint = ctx.accounts.mint.key().to_owned();
-        let mint_authority = ctx.accounts.mint.mint_authority.to_owned();
-
-        require!(
-            mint_authority.is_some(),
-            CcipRouterError::InvalidInputsPoolAccounts
-        );
-
-        let administrator = mint_authority.unwrap();
-
-        let token_admin_registry = &mut ctx.accounts.token_admin_registry;
-        token_admin_registry.version = 1;
-        token_admin_registry.administrator = administrator;
-        token_admin_registry.pending_administrator = Pubkey::new_from_array([0; 32]);
-        token_admin_registry.lookup_table = Pubkey::new_from_array([0; 32]);
-
-        emit!(AdministratorRegistered {
-            token_mint,
-            administrator,
-        });
-
-        Ok(())
+        instructions::v1::register_token_admin_registry_via_owner(ctx)
     }
 
     /// Sets the pool lookup table for a given token mint.
@@ -454,19 +347,7 @@ pub mod ccip_router {
         mint: Pubkey,
         pool_lookup_table: Pubkey,
     ) -> Result<()> {
-        let token_admin_registry = &mut ctx.accounts.token_admin_registry;
-        let previous_pool = token_admin_registry.lookup_table;
-        token_admin_registry.lookup_table = pool_lookup_table;
-
-        // TODO: Validate here that the lookup table has everything
-
-        emit!(PoolSet {
-            token: mint,
-            previous_pool_lookup_table: previous_pool,
-            new_pool_lookup_table: pool_lookup_table,
-        });
-
-        Ok(())
+        instructions::v1::set_pool(ctx, mint, pool_lookup_table)
     }
 
     /// Transfers the admin role of the token admin registry to a new admin.
@@ -483,21 +364,7 @@ pub mod ccip_router {
         mint: Pubkey,
         new_admin: Pubkey,
     ) -> Result<()> {
-        let token_admin_registry = &mut ctx.accounts.token_admin_registry;
-
-        if new_admin == Pubkey::new_from_array([0; 32]) {
-            token_admin_registry.pending_administrator = Pubkey::new_from_array([0; 32]);
-        } else {
-            token_admin_registry.pending_administrator = new_admin;
-        }
-
-        emit!(AdministratorTransferRequested {
-            token: mint,
-            current_admin: token_admin_registry.administrator,
-            new_admin,
-        });
-
-        Ok(())
+        instructions::v1::transfer_admin_role_token_admin_registry(ctx, mint, new_admin)
     }
 
     /// Accepts the admin role of the token admin registry.
@@ -512,17 +379,7 @@ pub mod ccip_router {
         ctx: Context<AcceptAdminRoleTokenAdminRegistry>,
         mint: Pubkey,
     ) -> Result<()> {
-        let token_admin_registry = &mut ctx.accounts.token_admin_registry;
-        let new_admin = token_admin_registry.pending_administrator;
-        token_admin_registry.administrator = new_admin;
-        token_admin_registry.pending_administrator = Pubkey::new_from_array([0; 32]);
-
-        emit!(AdministratorTransferred {
-            token: mint,
-            new_admin,
-        });
-
-        Ok(())
+        instructions::v1::accept_admin_role_token_admin_registry(ctx, mint)
     }
 
     /// Sets the token billing configuration.
@@ -1574,33 +1431,6 @@ fn build_receiver_discriminator_and_data(ramp_message: Any2SolanaMessage) -> Res
     data.extend_from_slice(&message);
 
     Ok(data)
-}
-
-fn validate_source_chain_config(
-    _source_chain_selector: u64,
-    _config: &SourceChainConfig,
-) -> Result<()> {
-    // As of now, the config has very few properties and there is nothing to validate yet.
-    // This is a placeholder to add validations as that config object grows.
-    Ok(())
-}
-
-fn validate_dest_chain_config(dest_chain_selector: u64, config: &DestChainConfig) -> Result<()> {
-    // TODO improve errors
-    require!(dest_chain_selector != 0, CcipRouterError::InvalidInputs);
-    require!(
-        config.default_tx_gas_limit != 0,
-        CcipRouterError::InvalidInputs
-    );
-    require!(
-        config.default_tx_gas_limit <= config.max_per_msg_gas_limit,
-        CcipRouterError::InvalidInputs
-    );
-    require!(
-        config.chain_family_selector != [0; 4],
-        CcipRouterError::InvalidInputs
-    );
-    Ok(())
 }
 
 #[error_code]
